minor refactors across many parts src/io: 	- added IO namespace 	- moved any input-output helper classes to the IO directory from 	  Utilities Toebox.cs: 	- fixed possible non-null reference warnings 	- refactored consructor src/suspension: 	- these classes got defined member variables 	- added default constructors
