+++ AutoChassis/src/io/PTSBuilder.cs
@@ -1,4 +1,6 @@
-namespace Utilities
+using Utilities;
+
+namespace IO
 {
     public static class PTSBuilder
     {
