+++ AutoChassis/src/utils/Point.cs
@@ -8,6 +8,13 @@ namespace Utilities
         public double y { get; set; }
         public double z { get; set; }
 
+        public Point()
+        {
+            x = 0;
+            y = 0;
+            z = 0;
+        }
+
         public Point(double x_, double y_, double z_ = 0)
         {
             x = x_;
